+++ plugins/modules/mongodb_maintenance.py
@@ -101,25 +101,8 @@ from ansible.module_utils.basic import AnsibleModule
 from ansible.module_utils.six import binary_type, text_type
 from ansible.module_utils.six.moves import configparser
 from ansible.module_utils._text import to_native
-from ansible_collections.community.mongodb.plugins.module_utils.mongodb_common import check_compatibility, missing_required_lib, load_mongocnf
+from ansible_collections.community.mongodb.plugins.module_utils.mongodb_common import check_compatibility, missing_required_lib, load_mongocnf, member_state
 from ansible_collections.community.mongodb.plugins.module_utils.mongodb_common import PyMongoVersion, PYMONGO_IMP_ERR, pymongo_found, MongoClient
-
-
-def member_state(client):
-    """Check if a replicaset exists.
-
-    Args:
-        client (cursor): Mongodb cursor on admin database.
-
-    Returns:
-        str: member state i.e. PRIMARY, SECONDARY
-    """
-    state = None
-    doc = client['admin'].command('replSetGetStatus')
-    for member in doc["members"]:
-        if "self" in member.keys():
-            state = str(member['stateStr'])
-    return state
 
 
 def put_in_maint_mode(client):
